// pdlearning/src/web-apps/src/libraries/domain-components/src/models/context-menu-action.model.ts
export enum ContextMenuAction {
  Archive = 'archive',
  Approve = 'approve',
  Reject = 'reject',
  Publish = 'publish',
  Unpublish = 'unpublish',
  Confirm = 'confirm',
  SendOffer = 'sendoffer',
  MoveToWaitlist = 'move-to-waitlist',
  Delete = 'delete',
  Duplicate = 'duplicate',
  Rename = 'rename',
  SubmitForApproval = 'submit-for-approval',
  CancellationRequest = 'cancellation-request',
  RescheduleRequest = 'reschedule-request',
  Present = 'present',
  Absent = 'absent',
  Verified = 'verified',
  Completed = 'completed',
  Incomplete = 'Incomplete',
  VerificationRejected = 'verificationRejected',
  TransferOwnership = 'transferOwnership',
  WithdrawManuallyByCA = 'withdrawManuallyByCA',
  ChangeRegistrationClass = 'ChangeRegistrationClass',
  ActivateCourseCriteria = 'ActivateCourseCriteria',
  DeActivateCourseCriteria = 'DeActivateCourseCriteria',
  ActivateCourseAutomate = 'ActivateCourseAutomate',
  DeActivateCourseAutomate = 'DeActivateCourseAutomate',
  Cancel = 'cancel',
  Send = 'send',
  Remove = 'remove',
  Reminder = 'reminder',
  NavigateToCSLCommunity = 'NavigateToCSLCommunity'
}